update OnboardingTemplate and add OnboardingItem in molecules
